pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Jarvis — personal AI assistant powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
jarvis = "jarvis.__main__:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
